projects/03_coffee_shop_full_stack/03_Francis_Sziebert_Coffee_Shop_Full_Stack/backend/requirements.txt
astroid
Click
ecdsa
Flask
Flask-SQLAlchemy
future
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
pycryptodome
pylint
python-jose-cryptodome
six
SQLAlchemy
typed-ast
Werkzeug
wrapt
Flask-Cors